backend/requirements.txt
# Core dependencies - Latest stable versions
fastapi
uvicorn
pydantic

# LangChain and related packages
langchain
langchain-core
langchain-google-genai
google-generativeai

# Utilities
python-dotenv
python-multipart
protobuf
tiktoken

# CrewAI and its dependencies
crewai
openai
litellm

# Ensure compatibility with crewai
pydantic-settings